backend/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-Cors==4.0.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
ldap3==2.9.1
MarkupSafe==2.1.5
pyasn1==0.5.1
python-dotenv==1.0.1
Werkzeug==3.0.1
zipp==3.17.0
